pyproject.toml
[project]
name = "tokenbank"
version = "0.1.0"
description = "TokenBank Phase 0 Private Agent Capacity Network bootstrap."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "jsonschema>=4.23.0",
  "pydantic>=2.10.0",
  "pyyaml>=6.0.2",
  "typer>=0.15.0",
  "uvicorn>=0.32.0",
]

[project.scripts]
tokenbank = "tokenbank.cli.main:main"

[dependency-groups]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tokenbank"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "."]
addopts = "-q"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
